Clean up makefile.
The intention was it might provide a reasonable interface for adding
packages to the toplevel.  It doesn't really, unfortunately.

js_of_ocaml must find cmi files for every module linked unless they are
expunged.  The expunge process currently says which modules to remove
rather than include.  This you would think would allow you to just
add packages but the .cmi files generally dont exist which means
you end up with massive EXPUNGE lists.

It will be more sensible for now to go back to listing modules to
include when expunging.

Either way the makefile is neater and the 2 main targets of a bare
toplevel and js+lwt toplevel are fine.